+++ src/app/medical-appoint-reg/medical-appointment-reg-page/medical-appointment-reg-page.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, ViewChild } from '@angular/core';
 import { Title } from '@angular/platform-browser';
 import { DataTransformService } from '../../shared/services/data-transform.service';
 import { SidebarMenuComponent } from '../../shared/sidebar-menu/sidebar-menu.component';
@@ -6,8 +6,7 @@ import { ToolbarComponent } from '../../shared/toolbar/toolbar.component';
 import { MatFormField, MatFormFieldModule } from '@angular/material/form-field';
 import { MatInputModule } from '@angular/material/input';
 import { MatSelectModule } from '@angular/material/select';
-import { MatDatepickerModule } from '@angular/material/datepicker';
-import { ErrorStateMatcher, MatNativeDateModule } from '@angular/material/core';
+import { ErrorStateMatcher } from '@angular/material/core';
 import { MatButtonModule, MatButton } from '@angular/material/button';
 import { FormBuilder, FormControl, FormGroupDirective, NgForm, ReactiveFormsModule, Validators } from '@angular/forms';
 import { CommonModule } from '@angular/common';
@@ -18,8 +17,8 @@ import { Observable, map, startWith } from 'rxjs';
 import { MatAutocompleteModule } from '@angular/material/autocomplete';
 import { ActivatedRoute, Router } from '@angular/router';
 import moment from 'moment';
-import { DateAdapter, MAT_DATE_FORMATS, MAT_DATE_LOCALE } from '@angular/material/core';
-import { CustomDateAdapter } from '../../shared/CustomDateAdapter';
+import { DialogComponent } from '../../shared/dialog/dialog.component';
+import { ConfirmDialogComponent } from '../../shared/confirm-dialog/confirm-dialog.component';
 
 export class MyErrorStateMatcher implements ErrorStateMatcher {
   isErrorState(control: FormControl | null, form: FormGroupDirective | NgForm | null): boolean {
@@ -31,19 +30,8 @@ export class MyErrorStateMatcher implements ErrorStateMatcher {
 @Component({
   selector: 'app-medical-appointment-reg-page',
   standalone: true,
-  imports: [ToolbarComponent, SidebarMenuComponent, MatFormFieldModule, MatInputModule, MatSelectModule, MatFormField, MatDatepickerModule, MatNativeDateModule, MatButtonModule, MatButton, ReactiveFormsModule, CommonModule, NgxMaterialTimepickerModule, HttpClientModule, MatAutocompleteModule],
-  providers: [DataTransformService, DataService, {provide: MAT_DATE_LOCALE, useValue: 'pt-BR'}, { provide: DateAdapter, useClass: CustomDateAdapter }, { provide: MAT_DATE_FORMATS, useValue: {
-    parse: {
-        dateInput: {month: 'short', year: 'numeric', day: 'numeric'}
-    },
-    display: {
-        dateInput: 'input',
-        monthYearLabel: {year: 'numeric', month: 'short'},
-        dateA11yLabel: {year: 'numeric', month: 'long', day: 'numeric'},
-        monthYearA11yLabel: {year: 'numeric', month: 'long'},
-    }
-  }}
-],
+  imports: [ToolbarComponent, SidebarMenuComponent, MatFormFieldModule, MatInputModule, MatSelectModule, MatFormField, MatButtonModule, MatButton, ReactiveFormsModule, CommonModule, NgxMaterialTimepickerModule, HttpClientModule, MatAutocompleteModule, DialogComponent, ConfirmDialogComponent],
+  providers: [DataTransformService, DataService],
   templateUrl: './medical-appointment-reg-page.component.html',
   styleUrl: './medical-appointment-reg-page.component.scss'
 })
@@ -60,6 +48,9 @@ export class MedicalAppointmentRegPageComponent implements OnInit {
     this.isEditing = !!this.activatedRoute.snapshot.paramMap.get('id')
    }
 
+   @ViewChild(DialogComponent) dialog!: DialogComponent;
+   @ViewChild(ConfirmDialogComponent) confirmDialog!: ConfirmDialogComponent;
+
    matcher = new MyErrorStateMatcher()
   
   appointRegistration = this.fb.group({
@@ -75,7 +66,7 @@ export class MedicalAppointmentRegPageComponent implements OnInit {
 
   setDate(date: Date) {
     let d = new Date(date);
-    d.setHours(d.getHours() + 24);
+    d.setHours(d.getHours());
     return d;
   }
     
@@ -87,8 +78,8 @@ export class MedicalAppointmentRegPageComponent implements OnInit {
 
     if (this.appointmentId) {
       this.dataService.getData('appointments/' + this.appointmentId).subscribe(appointment => {
-        const consultDate = moment(appointment.consultDate, 'DD-MM-YYYY').toDate();
-        appointment.consultDate = consultDate;
+        appointment.consultDate = this.dataTransformService.transformDateForForm(appointment.consultDate);
+
         this.appointRegistration.patchValue(appointment);
       })
     } else {
@@ -132,8 +123,8 @@ export class MedicalAppointmentRegPageComponent implements OnInit {
     if (this.appointRegistration.valid) {
         
         const appointment = {
-          idPatient: this.appointRegistration.value.idPatient,
-          name: this.appointRegistration.value.name,
+          idPatient: this.appointRegistration.getRawValue().idPatient,
+          name: this.appointRegistration.getRawValue().name,
           reason: this.appointRegistration.value.reason,
           consultDate: this.dataTransformService.formatDate(this.appointRegistration.value.consultDate),
           consultTime: this.appointRegistration.value.consultTime,
@@ -160,7 +151,7 @@ export class MedicalAppointmentRegPageComponent implements OnInit {
   }
     });
   } else {
-    window.alert('Preencha todos os campos obrigatórios corretamente.')
+    this.dialog.openDialog('Preencha todos os campos obrigatórios corretamente.');
 
     }
   }  
@@ -169,14 +160,14 @@ export class MedicalAppointmentRegPageComponent implements OnInit {
     if (this.appointRegistration.valid) {
       const appointment = {
         id: this.appointmentId,
-        idPatient: this.appointRegistration.value.idPatient,
-          name: this.appointRegistration.value.name,
-          reason: this.appointRegistration.value.reason,
-          consultDate: moment(this.appointRegistration.value.consultDate).format('DD-MM-YYYY'),
-          consultTime: this.appointRegistration.value.consultTime,
-          problemDescrip: this.appointRegistration.value.problemDescrip,
-          prescMed: this.appointRegistration.value.prescMed,
-          dosagesPrec: this.appointRegistration.value.dosagesPrec,
+        idPatient: this.appointRegistration.getRawValue().idPatient,
+        name: this.appointRegistration.getRawValue().name,
+        reason: this.appointRegistration.value.reason,
+        consultDate: this.dataTransformService.formatDate(this.appointRegistration.value.consultDate),
+        consultTime: this.appointRegistration.value.consultTime,
+        problemDescrip: this.appointRegistration.value.problemDescrip,
+        prescMed: this.appointRegistration.value.prescMed,
+        dosagesPrec: this.appointRegistration.value.dosagesPrec,
       }
   
       this.dataService.editData('appointments', this.appointmentId, appointment).subscribe(() => {
@@ -190,23 +181,30 @@ export class MedicalAppointmentRegPageComponent implements OnInit {
   
       });
     } else {
-      window.alert('Preencha todos os campos obrigatórios corretamente.');
+      this.dialog.openDialog('Preencha todos os campos obrigatórios corretamente.');
     }
   }
 
   editAppoint(){
     this.appointRegistration.enable();
+
+    this.appointRegistration.get('idPatient')!.disable();
+    this.appointRegistration.get('name')!.disable();
+
     this.saveDisabled = false;
   }
 
   deleteAppoint(){
-    this.dataService.deleteData('appointments', this.appointmentId).subscribe(() => {
-      window.alert('O registro foi excluído.');
-      this.router.navigate(['/lista-prontuarios']);
+    this.confirmDialog.openDialog("Tem certeza que deseja excluir a consulta? Essa ação não pode ser desfeita.")
+    const subscription = this.confirmDialog.confirm.subscribe(result => {
+      if (result) {
+        this.dataService.deleteData('appointments', this.appointmentId).subscribe(() => {
+          this.router.navigate(['/lista-prontuarios']);
+          subscription.unsubscribe();
+        });
+      } else {
+        subscription.unsubscribe();
+      }
     });
   }
-  
 }
-  
-  
-
